+++ hybrid_pool_executor/workers/asyncio/__init__.py
@@ -1,17 +1,27 @@
 from hybrid_pool_executor.base import ModuleSpec
+from hybrid_pool_executor.workers.asyncio.manager import AsyncManager, AsyncManagerSpec
 from hybrid_pool_executor.workers.asyncio.worker import (
-    AsyncManager,
-    AsyncManagerSpec,
+    AsyncTask,
     AsyncWorker,
     AsyncWorkerSpec,
 )
 
 MODULE_SPEC = ModuleSpec(
     name="async",
-    manager_class=AsyncManager,
-    manager_spec_class=AsyncManagerSpec,
-    worker_class=AsyncWorker,
-    worker_spec_class=AsyncWorkerSpec,
+    manager_type=AsyncManager,
+    manager_spec_type=AsyncManagerSpec,
+    worker_type=AsyncWorker,
+    worker_spec_type=AsyncWorkerSpec,
     tags=frozenset({"async"}),
     enabled=True,
+)
+
+
+__all__ = (
+    "MODULE_SPEC",
+    "AsyncManager",
+    "AsyncManagerSpec",
+    "AsyncTask",
+    "AsyncWorker",
+    "AsyncWorkerSpec",
 )
