+++ hybrid_pool_executor/workers/thread/__init__.py
@@ -1,17 +1,27 @@
 from hybrid_pool_executor.base import ModuleSpec
+from hybrid_pool_executor.workers.thread.manager import ThreadManager, ThreadManagerSpec
 from hybrid_pool_executor.workers.thread.worker import (
-    ThreadManager,
-    ThreadManagerSpec,
+    ThreadTask,
     ThreadWorker,
     ThreadWorkerSpec,
 )
 
 MODULE_SPEC = ModuleSpec(
     name="thread",
-    manager_class=ThreadManager,
-    manager_spec_class=ThreadManagerSpec,
-    worker_class=ThreadWorker,
-    worker_spec_class=ThreadWorkerSpec,
+    manager_type=ThreadManager,
+    manager_spec_type=ThreadManagerSpec,
+    worker_type=ThreadWorker,
+    worker_spec_type=ThreadWorkerSpec,
     tags=frozenset({"thread", "async"}),
     enabled=True,
+)
+
+
+__all__ = (
+    "MODULE_SPEC",
+    "ThreadManager",
+    "ThreadManagerSpec",
+    "ThreadTask",
+    "ThreadWorker",
+    "ThreadWorkerSpec",
 )
