+++ tox.ini
@@ -2,14 +2,15 @@
 envlist =
     autoflake
     black
+    flake8
+    isort
+    mypy
     py38
     py38_extra
     py39
     py39_extra
     py310
     py310_extra
-    flake8
-    isort
 
 [testenv]
 deps =
@@ -33,6 +34,10 @@ commands = pflake8 hybrid_pool_executor tests setup.py {posargs}
 [testenv:isort]
 deps = isort
 commands = isort --check . {posargs}
+
+[testenv:mypy]
+deps = mypy
+commands = mypy hybrid_pool_executor tests
 
 [testenv:py38_extra]
 extras = extra
